requirements.txt
# Core dependencies
behave==1.2.6
playwright==1.55.0
PyYAML==6.0.3

# Optional: colorized console output
colorama==0.4.6

# (Playwright uses these internally, safe to pin)
pyee==13.0.0
typing_extensions==4.15.0

# Reporting
behave-html-formatter==0.9.10